Ajout chargement data.sql et modification prix et VAT en int

+++ src/main/java/fr/eservices/drive/model/Article.java
@@ -8,8 +8,8 @@ import java.util.List;
 public class Article {
 	@Id
 	private String ean13;
-	private double price;
-	private double vat;
+	private int price;
+	private int vat;
 	private String name;
 	private String img;
 
@@ -24,19 +24,19 @@ public class Article {
 		this.ean13 = ean13;
 	}
 
-	public double getPrice() {
+	public int getPrice() {
 		return price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
 
-	public double getVat() {
+	public int getVat() {
 		return vat;
 	}
 
-	public void setVat(double tva) {
+	public void setVat(int tva) {
 		this.vat = tva;
 	}
